part1_core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rdkit-core"
version = "0.1.0"
description = "Shared contracts and tool layer for the R&D Multi-Agent ML System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0,<3.0",
    "mlflow>=2.10,<3.0",
    "pandas>=2.0,<3.0",
    "docker>=7.0,<8.0",
    "fsspec>=2024.0",
    "openai>=1.20,<2.0",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "tenacity>=8.0,<10.0",
]

[project.optional-dependencies]
s3 = ["boto3>=1.34", "s3fs>=2024.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "pandas-stubs>=2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/rdkit_core"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
